New cmdlets changes and fixes

+++ EWSContacts/Module/functions/service/Connect-EXCExchange.ps1
@@ -30,22 +30,33 @@
 	{
 		## Load Managed API dll  
 		###CHECK FOR EWS MANAGED API, IF PRESENT IMPORT THE HIGHEST VERSION EWS DLL, ELSE EXIT
-		$EWSDLL = (($(Get-ItemProperty -ErrorAction SilentlyContinue -Path Registry::$(Get-ChildItem -ErrorAction SilentlyContinue -Path 'Registry::HKEY_LOCAL_MACHINE\SOFTWARE\Microsoft\Exchange\Web Services' | Sort-Object Name -Descending | Select-Object -First 1 -ExpandProperty Name)).'Install Directory') + "Microsoft.Exchange.WebServices.dll")
-		if (Test-Path $EWSDLL)
-		{
-			Import-Module $EWSDLL
-		}
-		else
-		{
-			"$(get-date -format yyyyMMddHHmmss):"
-			"This script requires the EWS Managed API 1.2 or later."
-			"Please download and install the current version of the EWS Managed API from"
-			"http://go.microsoft.com/fwlink/?LinkId=255472"
-			""
-			"Exiting Script."
-			$exception = New-Object System.Exception ("Managed Api missing")
-			throw $exception
-		}
+        if (Test-Path ($script:ModuleRoot + "/bin/Microsoft.Exchange.WebServices.dll")) {
+            Import-Module ($script:ModuleRoot + "/bin/Microsoft.Exchange.WebServices.dll")
+            $Script:EWSDLL = $script:ModuleRoot + "/bin/Microsoft.Exchange.WebServices.dll"
+            write-verbose ("Using EWS dll from Local Directory")
+        }
+        else {
+
+			
+            ## Load Managed API dll  
+            ###CHECK FOR EWS MANAGED API, IF PRESENT IMPORT THE HIGHEST VERSION EWS DLL, ELSE EXIT
+            $EWSDLL = (($(Get-ItemProperty -ErrorAction SilentlyContinue -Path Registry::$(Get-ChildItem -ErrorAction SilentlyContinue -Path 'Registry::HKEY_LOCAL_MACHINE\SOFTWARE\Microsoft\Exchange\Web Services'|Sort-Object Name -Descending| Select-Object -First 1 -ExpandProperty Name)).'Install Directory') + "Microsoft.Exchange.WebServices.dll")
+            if (Test-Path $EWSDLL) {
+                Import-Module $EWSDLL
+                $Script:EWSDLL = $EWSDLL 
+            }
+            else {
+                "$(get-date -format yyyyMMddHHmmss):"
+                "This script requires the EWS Managed API 1.2 or later."
+                "Please download and install the current version of the EWS Managed API from"
+                "http://go.microsoft.com/fwlink/?LinkId=255472"
+                ""
+                "Exiting Script."
+                exit
+
+
+            } 
+        }
 		
 		## Set Exchange Version  
 		$ExchangeVersion = [Microsoft.Exchange.WebServices.Data.ExchangeVersion]::Exchange2010_SP2
